requirements.txt
feedparser
python-dateutil
requests
sendgrid
